% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_vcfR.R
\name{calc_missing_gt}
\alias{calc_missing_gt}
\title{Calculate true number of missing genotypes in a vcfR object}
\usage{
calc_missing_gt(x, pct = FALSE)
}
\arguments{
\item{x}{A vcfR object}

\item{pct}{Return value as a percent (out of 100) instead of a proportion (default)}
}
\value{
Proportion of missing genotypes
}
\description{
Calculate true number of missing genotypes in a vcfR object
}
\examples{
calc_missing_gt(vcf_example)
}
\keyword{genotypes}
\keyword{missingness}
\keyword{vcf}
