% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_popgen_tools.R
\name{dapc_workflow}
\alias{dapc_workflow}
\title{Make DAPC plots}
\usage{
dapc_workflow(x, formula, name)
}
\arguments{
\item{x}{A genclone object}

\item{formula}{Hierarchical formula to group the populations by in plots}

\item{name}{Used to label plot axes and titles}
}
\value{
Three DAPC plots
}
\description{
Make DAPC plots
}
\examples{
dapc_workflow(ultra_filtered_genclone, ~Nursery, "Nurseries as Pops")
}
\keyword{dapc}
\keyword{genclone}
\keyword{genind}
