% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_vcfR.R
\name{missingGT_maf_filter}
\alias{missingGT_maf_filter}
\title{Filter VCF based on number of minor alleles in dataset}
\usage{
missingGT_maf_filter(x, min_count = 2, isGTmissing = TRUE)
}
\arguments{
\item{x}{A vcfR object}

\item{min_count}{Minimum number of times a variant's minor allele must be called to avoid being removed}

\item{isGTmissing}{Remove all variants with missing data (default) or not (FALSE).}
}
\value{
VCF filtered by minor allele count and by missingness
}
\description{
Filter VCF based on number of minor alleles in dataset
}
\examples{
missingGT_maf_filter(vcf_example)
}
\keyword{MAF}
\keyword{filter}
\keyword{missingness}
\keyword{vcf}
