% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_popgen_tools.R
\name{pca_workflow}
\alias{pca_workflow}
\title{Make PCA plots}
\usage{
pca_workflow(x, pop_def, pop_map, filter_strata = TRUE)
}
\arguments{
\item{x}{A genlight object}

\item{pop_def}{Define what populations are to be used}

\item{pop_map}{Reference population map/strata}

\item{filter_strata}{Filter down your strata to only samples used}
}
\value{
PCA plots
}
\description{
Make PCA plots
}
\examples{
pca_workflow(ultra_filtered_genlight, 'Nursery', strata_defs)
}
\keyword{genlight}
\keyword{pca}
