% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_vcfR.R
\name{plot_violins}
\alias{plot_violins}
\title{Make violin plots for all samples in a vcfR object}
\usage{
plot_violins(x, samples_per_row = 10, fill_group_aesthetic = NULL)
}
\arguments{
\item{x}{A vcfR object}

\item{samples_per_row}{Number of violins per row}

\item{fill_group_aesthetic}{Fills violin plots based on a vcf statistic. Currently not supported.}
}
\value{
Violin plots of all samples
}
\description{
Make violin plots for all samples in a vcfR object
}
\examples{
plot_violins(ultra_filtered_vcf)
}
\keyword{filter}
\keyword{vcf}
