% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_vcfR.R
\name{samples_filter}
\alias{samples_filter}
\title{Filter out samples of a vcfR object based on depth}
\usage{
samples_filter(x, lower_quant = 0.05, stat = "median", filter_bot = TRUE,
  filter_top = FALSE, remove_nonpolymorphs = TRUE)
}
\arguments{
\item{x}{A vcfR object}

\item{lower_quant}{lower quantile to filter by. The top quantile will be \code{1-lower_quant}}

\item{stat}{Statistic to filter by. Right now, supports "median" to filter on median depth, "mean" to filter on mean depth, or "missingness" to filter on percent missing data.}

\item{filter_bot}{Omit samples in the bottom quantile}

\item{filter_top}{Omit samples in the top quantile}

\item{remove_nonpolymorphs}{After omitting samples, some variants will be nonpolymorphic. These should be removed.}
}
\value{
A filtered vcfR object.
}
\description{
Filter out samples of a vcfR object based on depth
}
\examples{
ultrafiltered_vcf <- samples_filter(filtered_vcf)
}
\keyword{filter}
\keyword{vcf}
