% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.experiment.R
\name{set.experiment}
\alias{set.experiment}
\title{Select the experiment where the results would be saved}
\usage{
set.experiment(name, overwrite = F, results.path = "./results/")
}
\arguments{
\item{x}{Number of threads used for the calculations. If it is NA}
}
\value{
Returns a vector with the same size of the input vector y with the
}
\description{
This is the main function of the miRNAss package and implements the miRNA
}
\examples{
# First construct the label vector with the CLASS column
a = 1+1

}
