% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tic.R
\name{tic}
\alias{tic}
\title{Select the experiment where the results would be saved}
\usage{
tic(name = "")
}
\arguments{
\item{x}{Number of threads used for the calculations. If it is NA}
}
\value{
Returns a vector with the same size of the input vector y with the
}
\description{
This is the main function of the miRNAss package and implements the miRNA
}
\examples{
# First construct the label vector with the CLASS column
a = 1+1

}
