% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_popgen_tools.R
\name{tree_workflow}
\alias{tree_workflow}
\title{Make phylogenetic trees colored according to strata}
\usage{
tree_workflow(
  x,
  formula = NULL,
  name,
  scale_quality = "population",
  vcf = NULL,
  geo_tree = NULL,
  treeType = "upgma",
  dist = "provesti.dist",
  input_type = "genclone"
)
}
\arguments{
\item{x}{Preferably a genclone object, genlights sometimes break it. Must have strata}

\item{formula}{Hierarchical formula to group the populations by in plots. Do not wrap in quotes... I think.}

\item{name}{Used to label plot axes and titles.}

\item{scale_quality}{What to color the tree tips by. Use 'population' to color by any "formula" you set. Other options are: 'missing', 'median_depth', 'mean_depth', 'heterozygosity', and 'allele_depth'}

\item{vcf}{Original vcfR object, used to calculate statistics on. Be sure it's the same exact individuals and variants as the data in `x`.}

\item{geo_tree}{Optionally supply your own tree instead of running poppr::aboot}

\item{treeType}{Algorithm to make tree by. Make sure the package to use is loaded.}

\item{dist}{Algorithm to calculate distance. Make sure the package to use is loaded.}

\item{input_type}{Specify if a genclone or genlight is inputted}
}
\value{
Phylogenetic tree plot
}
\description{
Make phylogenetic trees colored according to strata
}
\examples{
tree_workflow(ultra_filtered_genlight, ~Nursery, "Nursery of origin", "population", ultra_filtered_vcf)
}
\keyword{ape}
\keyword{genclone}
\keyword{genind}
\keyword{phylogeny}
